Rename file and make utils folder
