+++ static/app/views/dashboards/datasetConfig/utils/getSeriesRequestData.spec.tsx
@@ -4,7 +4,7 @@ import {WidgetFixture} from 'sentry-fixture/widget';
 import {WidgetQueryFixture} from 'sentry-fixture/widgetQuery';
 
 import {DiscoverDatasets} from 'sentry/utils/discover/types';
-import {getSeriesRequestData} from 'sentry/views/dashboards/datasetConfig/utils';
+import {getSeriesRequestData} from 'sentry/views/dashboards/datasetConfig/utils/getSeriesRequestData';
 import {DisplayType} from 'sentry/views/dashboards/types';
 
 describe('utils', () => {
